Reorganize ComponentsList components
